refactor: add createStyled
This is a relatively pure refactor without new functionality, in
preparation for adding `styled.extend()`

First, add `createStyled` to emit `styled` given a generator function.

Second, replace `styledWithGenerator` by `styled.*.withGenerator`
(styled-components) and `styled(*, { generator })` (emotion). This
avoids the creator needing to emit more than one function with different
capabitilies.
